Add `rake users:initialize_admin` task
When we push up to Heroku, we'll need to initialize the admin user. This
task lets us do that with environment variables.

The second big thing in this commit is switching to an injection
strategy for testing classes that depend on ENV (vs. stubbing or using
the "magic" `.env.test`)

Major Changes:

  - Add: `users:initialize_admin` rake task
  - Add: `AdminInitializer` helper class + specs
  - Add: README instructions for initializing admin in development

  - Update: Instead of stubbing ENV, inject it (affects AdminInitializer
            and AmazonProductAPI::HTTPClient)
  - Remove: `.env.test` (in favor of injecting)

Note:

I'm assuming we'll only want one "root" admin user. If we want to add
multiple admins, we'll want to change this task to take arguments or
user input or something.